requirements.txt
asgiref==3.8.1
channels==4.1.0
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gevent==24.2.1
greenlet==3.0.3
PyJWT==2.8.0
sqlparse==0.5.0
zope.event==5.0
zope.interface==6.3
